api/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy==2.0.23
alembic==1.13.1
psycopg2-binary
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
pydantic[email]